requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
pandas
python-multipart
